# Frontend/unifortrade/openapi.yaml
openapi: 3.1.0
info:
  title: UniforTrade API
  description: |
    Marketplace de economia circular do campus — estudantes anunciam itens (livros, calculadoras, componentes eletrônicos, jalecos, móveis etc.) para venda ou doação a outros estudantes.

    Fluxo de uma transação: uma listing ACTIVE é reivindicada (`POST /{id}/claim`, vira CLAIMED), as duas partes trocam contato (`GET /{id}/contact`) para combinar local e horário, e o vendedor confirma a entrega presencial (`POST /{id}/complete`). Qualquer uma das partes pode desistir antes da entrega (`DELETE /{id}/claim`), o que devolve a listing para ACTIVE.

    A maioria dos endpoints de leitura é pública; escrita e ações de fluxo exigem um usuário autenticado, e endpoints de categoria (além de GET) exigem a role ADMIN.

    Autenticação: o cadastro é `POST /api/users`, que cria a conta e retorna o perfil, mas não devolve tokens. Faça login em `POST /auth/login` para receber o par `accessToken` / `refreshToken`, renove em `POST /auth/refresh` e encerre a sessão no cliente com `POST /auth/logout`. Use o `accessToken` no botão "Authorize" acima para enviá-lo como `Authorization: Bearer <token>`.
  version: 0.0.1
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
  /auth/refresh:
    post:
      tags:
      - auth-controller
      operationId: refresh
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/logout:
    post:
      tags:
      - auth-controller
      operationId: logout
      responses:
        "200":
          description: OK
  /auth/login:
    post:
      tags:
      - auth-controller
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /api/users:
    post:
      tags:
      - user-controller
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CurrentUserResponse"
  /api/users/{userId}/ratings:
    get:
      tags:
      - rating-controller
      operationId: findByUserId
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserRatingsResponse"
    post:
      tags:
      - rating-controller
      operationId: create_1
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRatingRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RatingResponse"
  /api/users/me/notifications/{id}/dismiss:
    post:
      tags:
      - notification-controller
      operationId: dismiss
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/NotificationResponse"
  /api/listings:
    get:
      tags:
      - listing-controller
      operationId: search
      parameters:
      - name: filters
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/ListingSearchRequest"
      - name: pageable
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageResponseListingSummaryResponse"
    post:
      tags:
      - listing-controller
      operationId: create_2
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateListingRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ListingResponse"
  /api/listings/{id}/complete:
    post:
      tags:
      - listing-controller
      operationId: complete
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ListingResponse"
  /api/listings/{id}/claim:
    post:
      tags:
      - listing-controller
      operationId: claim
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ListingResponse"
    delete:
      tags:
      - listing-controller
      operationId: cancelClaim
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ListingResponse"
  /api/categories:
    get:
      tags:
      - category-controller
      operationId: findAllByOrderByNameAsc
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryResponse"
    post:
      tags:
      - category-controller
      operationId: create_3
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CategoryResponse"
  /api/users/{id}:
    get:
      tags:
      - user-controller
      operationId: findById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProfileResponse"
    delete:
      tags:
      - user-controller
      operationId: delete
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
    patch:
      tags:
      - user-controller
      operationId: patch
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchUserRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CurrentUserResponse"
  /api/users/me:
    get:
      tags:
      - self-controller
      operationId: getSelf
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CurrentUserResponse"
    delete:
      tags:
      - self-controller
      operationId: deleteSelf
      responses:
        "200":
          description: OK
    patch:
      tags:
      - self-controller
      operationId: patchSelf
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchUserRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CurrentUserResponse"
  /api/listings/{id}:
    get:
      tags:
      - listing-controller
      operationId: findById_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ListingResponse"
    delete:
      tags:
      - listing-controller
      operationId: delete_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
    patch:
      tags:
      - listing-controller
      operationId: patch_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchListingRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ListingResponse"
  /api/users/uniforId/{uniforId}:
    get:
      tags:
      - user-controller
      operationId: findByUniforId
      parameters:
      - name: uniforId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProfileResponse"
  /api/users/me/notifications:
    get:
      tags:
      - notification-controller
      operationId: findAll
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NotificationResponse"
  /api/stats:
    get:
      tags:
      - stats-controller
      operationId: getStats
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/StatsResponse"
  /api/listings/{id}/contact:
    get:
      tags:
      - listing-controller
      operationId: getContact
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ContactResponse"
  /api/categories/{id}:
    get:
      tags:
      - category-controller
      operationId: findById_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CategoryResponse"
    delete:
      tags:
      - category-controller
      operationId: delete_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
components:
  schemas:
    RefreshRequest:
      type: object
      properties:
        refreshToken:
          type: string
          minLength: 1
      required:
      - refreshToken
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        tokenType:
          type: string
    LoginRequest:
      type: object
      properties:
        uniforId:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1
      required:
      - password
      - uniforId
    CreateUserRequest:
      type: object
      properties:
        uniforId:
          type: string
        password:
          type: string
        name:
          type: string
        bio:
          type: string
        profilePicture:
          type: string
        phoneNumber:
          type: string
        email:
          type: string
    ContactInfoResponse:
      type: object
      properties:
        phoneNumber:
          type: string
        email:
          type: string
    CurrentUserResponse:
      type: object
      properties:
        uniforId:
          type: string
        profile:
          $ref: "#/components/schemas/ProfileResponse"
        contactInfo:
          $ref: "#/components/schemas/ContactInfoResponse"
    ProfileResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    CreateRatingRequest:
      type: object
      properties:
        score:
          type: integer
          format: int32
        comment:
          type: string
    RatingResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        raterId:
          type: integer
          format: int64
        raterDisplayName:
          type: string
        ratedId:
          type: integer
          format: int64
        score:
          type: integer
          format: int32
        comment:
          type: string
        createdAt:
          type: string
          format: date-time
    NotificationResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        message:
          type: string
        dismissed:
          type: boolean
        createdAt:
          type: string
          format: date-time
        linkType:
          type: string
          enum:
          - LISTING
          - USER
        linkId:
          type: integer
          format: int64
    CreateListingRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        price:
          type: number
        gallery:
          type: array
          items:
            type: string
        categoryId:
          type: integer
          format: int64
    CategoryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        slug:
          type: string
        imageUrl:
          type: string
    ListingResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum:
          - SALE
          - DONATION
        price:
          type: number
        gallery:
          type: array
          items:
            type: string
        status:
          type: string
          enum:
          - ACTIVE
          - CLAIMED
          - COMPLETED
        category:
          $ref: "#/components/schemas/CategoryResponse"
        owner:
          $ref: "#/components/schemas/ProfileResponse"
        createdAt:
          type: string
          format: date-time
    CreateCategoryRequest:
      type: object
      properties:
        name:
          type: string
        imageUrl:
          type: string
    PatchUserRequest:
      type: object
      properties:
        name:
          type: string
        bio:
          type: string
        profilePicture:
          type: string
        phoneNumber:
          type: string
        email:
          type: string
    PatchListingRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        price:
          type: number
        gallery:
          type: array
          items:
            type: string
        categoryId:
          type: integer
          format: int64
    UserRatingsResponse:
      type: object
      properties:
        ratings:
          type: array
          items:
            $ref: "#/components/schemas/RatingResponse"
        average:
          type: number
          format: double
        count:
          type: integer
          format: int64
    StatsResponse:
      type: object
      properties:
        completedClaimsCount:
          type: integer
          format: int64
        completedSaleClaimsCount:
          type: integer
          format: int64
        completedDonationClaimsCount:
          type: integer
          format: int64
        registeredUsersCount:
          type: integer
          format: int64
    ListingSearchRequest:
      type: object
      properties:
        keyword:
          type: string
        categoryId:
          type: integer
          format: int64
        ownerId:
          type: integer
          format: int64
        type:
          type: string
          enum:
          - SALE
          - DONATION
        status:
          type: string
          enum:
          - ACTIVE
          - CLAIMED
          - COMPLETED
        minPrice:
          type: number
        maxPrice:
          type: number
    Pageable:
      type: object
      properties:
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
        sort:
          type: array
          items:
            type: string
    ListingSummaryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        price:
          type: number
        thumbnailUrl:
          type: string
        status:
          type: string
          enum:
          - ACTIVE
          - CLAIMED
          - COMPLETED
        categoryName:
          type: string
    PageResponseListingSummaryResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/ListingSummaryResponse"
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        last:
          type: boolean
    ContactResponse:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        displayName:
          type: string
        email:
          type: string
        phone:
          type: string
  securitySchemes:
    bearerAuth:
      type: http
      description: Access token returned by POST /auth/login or POST /auth/refresh.
      scheme: bearer
      bearerFormat: JWT